Report the last communication date of Nitoo and Zigbee devices

+++ src/OpenNetty/OpenNettyDeviceIdentifier.cs
@@ -147,7 +147,7 @@ public readonly struct OpenNettyDeviceIdentifier : IEquatable<OpenNettyDeviceIde
     /// </summary>
     /// <param name="identifier">The serial number.</param>
     /// <returns>A new <see cref="OpenNettyDeviceIdentifier"/> instance representing a Zigbee serial number.</returns>
-    public static OpenNettyDeviceIdentifier FromZigbeeSerialNumber(ulong identifier)
+    public static OpenNettyDeviceIdentifier FromZigbeeSerialNumber(uint identifier)
         => new(OpenNettyDeviceIdentifierType.ZigbeeSerialNumber, identifier.ToString("X8", CultureInfo.InvariantCulture));
 
     /// <summary>
